requirements.txt
torch>=2.3.0
transformers>=4.45.0
datasets>=2.20.0
peft>=0.12.0
accelerate>=0.33.0
deepspeed>=0.15.0
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.5.0
pandas>=2.2.0
matplotlib>=3.9.0
pyyaml>=6.0.0
tqdm>=4.66.0
wandb>=0.17.0
